Added package, Python 2 compatibility and support for adding to <head>
The source code was moved into a package, and the dependencies were moved
into the package as subpackages, which means `sys.path` is no longer being
modified.

With the help of the `six` module and some other adjustments (for example,
using `__future__`), Python 2.6 and up should now be supported, and
the `retro` branch has been deleted.

Drafts may now specify one or more values for the `head` metadata variable, and
they will be output in the template wherever `{{ head }}` is found, with each
value on its own line. This allows you to include extra CSS or JavaScript in
the resulting page's <head> element.