Added -c option specifying location of presto.ini
Presto's current working directory no longer has to be the same as the one
containing presto.ini; add -c/--config to specify where the presto.ini
is located.

Special note: in presto.ini, relative paths are now computed relative
to the location of presto.ini, not the current working directory.